pyproject.toml
[tool.poetry]
name = "zwift-click-handler"
version = "0.1.0"
description = ""
authors = ["Joshua Taillon <jat255@gmail.com>"]
readme = "README.md"
packages = [{include = "zwift_click_handler"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.1"
bleak = "^0.22.1"
rich = "^13.7.1"
cryptography = "^42.0.7"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
